adding game and memoria games
Two games called ‘game’ and ‘memoria’ were added. The first one is a copy of Plants Vs Zombies, and in this case, you can customize who is the plant and who is the zombie by uploading your own pictures. The game called ‘memoria’ is a memory game where the user selects 2 cards, they will show their images and then return to their initial position. The user needs to match pairs of images.